[CofenseTriage] Rename module to CofenseTriagev2
on-behalf-of: @Cofense <oss@cofense.com>
