+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/CofenseTriage.yml
