+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/CofenseTriage_description.md
