+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/test/conftest.py
@@ -7,8 +7,8 @@ demisto_params = {
     "user": "user",
 }
 with patch("demistomock.params", lambda: demisto_params):
-    from CofenseTriage2 import CofenseTriage  # noqa: 401 - this is used in other test files
-    from CofenseTriage2.CofenseTriage import TriageInstance  # noqa: 401 - this is used in other test files
+    from CofenseTriagev2 import CofenseTriage  # noqa: 401 - this is used in other test files
+    from CofenseTriagev2.CofenseTriage import TriageInstance  # noqa: 401 - this is used in other test files
 
 
 @pytest.fixture
