+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/test/fixtures/malformed_json.not_json
