+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/test/fixtures/threat_indicators.json
