+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/test/test_triage_reporter.py
@@ -1,4 +1,4 @@
-from CofenseTriage2.CofenseTriage import TriageReporter
+from CofenseTriagev2.CofenseTriage import TriageReporter
 
 
 class TestTriageReporter:
